compliance_dashboard/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
reportlab==4.0.7
openpyxl==3.1.2
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
